feat: rename and add packages in loader
