+++ loader/app/Server.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, Request
 
-from loader.src.cloud_services.cloud_service import CloudService
+from loader.app.cloud_services.cloud_service import CloudService
 
 
 class Server:
