--- loader/src/objects/json_object.py
@@ -1,11 +1,0 @@
-import json
-
-from loader.src.objects.object import Object
-
-class JsonObject(Object):
-    def __init__(self, content : json):
-        self.content = content
-        self.mimtype = "json"
-
-    def get_formated(self):
-        return json.dumps(self.content)
