--- loader/src/objects/object.py
@@ -1,5 +1,0 @@
-class Object:
-    def get_formated(content : any):
-        pass;
-
-
