+++ loader/tests/test_load.py
@@ -1,0 +1,4 @@
+# récupération du document avec un url fictif
+# load dans un faux micros service mocké
+# test cas d'erreurs
+# retourner lien ou erreur
